testfind: fake dird-fd communication when sending filesets
- Faking how sockets work to mimick the real interaction between 
dird and fd
- Renamed `testfind_jcr` to `testfind_fd`
- Moved testfind from dird to tools
- Enabled joblog messages
asd
