+++ core/src/tools/testfind_fd.h
@@ -23,19 +23,13 @@
 #define TESTFIND_JCR_H
 
 #include "include/jcr.h"
-#include "findlib/find.h"
 #include "dird/dird_conf.h"
+#include "findlib/find.h"
 
-void launchFileDaemonLogic(directordaemon::FilesetResource* jcr_fileset,
-                           const char* configfile,
-                           bool print_attrs);
-void SetOptions(findFOPTS* fo, const char* opts);
-bool setupFileset(FindFilesPacket* ff,
-                  directordaemon::FilesetResource* jcr_fileset);
-void CountFiles(FindFilesPacket* ar);
+void ProcessFileset(directordaemon::FilesetResource* director_fileset,
+                    const char* configfile,
+                    bool print_attrs);
+void UpdateFilestats(FindFilesPacket* ffp);
 int PrintFile(JobControlRecord*, FindFilesPacket* ff, bool);
-int testfindLogic(JobControlRecord* jcr,
-                  FindFilesPacket* ff_pkt,
-                  bool top_level);
 
 #endif  // TESTFIND_JCR_H
