+++ systemtests/tests/testfind/etc/bareos/bareos-fd.d/messages/Standard.conf.in
@@ -2,4 +2,5 @@ Messages {
   Name = Standard
   Director = bareos-dir = all, !skipped, !restored
   Description = "Send relevant messages to the Director."
+  append = "@logdir@/bareos.log" = all, !skipped, !saved, !audit
 }
